added a Prometheus metric for expired messages; moved test
