+++ server/sms/nexmo_sms_sender_test.go
@@ -1,8 +1,6 @@
 package sms
 
 import (
-	"expvar"
-	"net/http"
 	"testing"
 	"time"
 
@@ -36,31 +34,4 @@ func TestNexmoSender_SendWithError(t *testing.T) {
 	time.Sleep(3 * timeInterval)
 	a.Error(err)
 	a.Equal(ErrRetryFailed, err)
-}
-
-func TestNexmoSender_SendExpiredMessage(t *testing.T) {
-	a := assert.New(t)
-
-	port := createRandomPort(7000, 8000)
-	URL = "http://127.0.0.1" + port
-
-	sender := createNexmoSender(t)
-	// no request should be made in case the sms is expired
-	go dummyNexmoEndpointWithHandlerFunc(t, nil, port, func(t *testing.T, countCh chan bool) http.HandlerFunc {
-		return func(w http.ResponseWriter, r *http.Request) {
-			a.FailNow("Nexmo call not expected.")
-		}
-	})
-
-	msg := encodeProtocolMessage(t, 0)
-	expires := time.Now().Add(-1 * time.Hour)
-	msg.Expires = &expires
-
-	err := sender.Send(&msg)
-	time.Sleep(3 * timeInterval)
-	a.NoError(err)
-
-	expectedExpired := expvar.NewInt("total_expired_messages")
-	expectedExpired.Add(1)
-	a.Equal(expectedExpired, mTotalExpiredMessages)
 }
